Code cleanup
  * fldigi.cxx/fldigi.h
    - Changed scope of variables to static if not accessed
      outside of fl_digi.cxx
    - added gettext precompiler notation where needed
  * Removed unused variables / code
    - src/compat/getsysinfo.c
    - src/widgets/Fl_Text_Display_mod_1_3.cxx
    - src/xmlrpcpp/XmlRpcSource.cpp

+++ src/xmlrpcpp/XmlRpcSource.cpp
@@ -189,7 +189,7 @@ namespace XmlRpc {
       _ssl = 0;
     }
 #endif
-    if (_fd != -1)
+    if ( (int)_fd != -1)
     {
       XmlRpcUtil::log(2,"XmlRpcSource::close: closing socket %d.", _fd);
       XmlRpcSocket::close(_fd);
